fix all imports & move all managers into the new resources folder
