+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_store_app/app/main_app.dart';
+
+import 'app/main_app.dart';
 
 void main() {
   runApp(MainApp());
