+++ lib/resources/color_manager.dart
