+++ lib/resources/style_manager.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_store_app/presentation/color_manager.dart';
-import 'package:flutter_store_app/presentation/fonts_manager.dart';
+import 'color_manager.dart';
+import 'fonts_manager.dart';
 
 class TextStyleManager {
   static const defaultFontFamily = AppFontFamilies.montserrat;
