+++ lib/resources/values_manager.dart
